Renamed and moved App/Place to Repository/ApiRequest
